pyproject.toml
[project]
name = "mcp-with-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.6",
    "langsmith>=0.4.25",
    "mcp[cli]>=1.13.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
]
